Split examples and regressions tests. Examples must pass. Regressions are xfail (for now)
